Move StructureProperties and StructureGeoJSON to structures dux module
